+++ Backend/Services/LuaServiceLib/IRacingMethodCollection.cs
@@ -32,6 +32,7 @@ function iracing_send_track_info(); iracing:send_track_info(); end
 function iracing_send_weather_info(); iracing:send_weather_info(); end
 function iracing_send_current_session(); iracing:send_current_session(); end
 function iracing_send_session_state(); iracing:send_session_state(); end
+function iracing_send_race_flags(); iracing:send_race_flags(); end
 ");
             }
 
@@ -63,6 +64,12 @@ function iracing_send_session_state(); iracing:send_session_state(); end
             public void send_session_state()
             {
                 EventBus.PublishEvent(EventFactory.CreateIRacingCommandSendSessionState());
+            }
+
+            [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "This is expose in Lua, so we want to keep that naming style")]
+            public void send_race_flags()
+            {
+                EventBus.PublishEvent(EventFactory.CreateIRacingCommandSendRaceFlags());
             }
         }
     }