+++ Shared/EventHandler.cs
@@ -143,6 +143,9 @@ namespace Slipstream.Shared
         public delegate void OnIRacingCommandSendSessionStateHandler(EventHandler source, EventHandlerArgs<Shared.Events.IRacing.IRacingCommandSendSessionState> e);
         public event OnIRacingCommandSendSessionStateHandler? OnIRacingCommandSendSessionState;
 
+        public delegate void OnIRacingCommandSendRaceFlagsHandler(EventHandler source, EventHandlerArgs<Shared.Events.IRacing.IRacingCommandSendRaceFlags> e);
+        public event OnIRacingCommandSendRaceFlagsHandler? OnIRacingCommandSendRaceFlags;
+
         public delegate void OnIRacingDriverIncidentHandler(EventHandler source, EventHandlerArgs<Shared.Events.IRacing.IRacingDriverIncident> e);
         public event OnIRacingDriverIncidentHandler? OnIRacingDriverIncident;
         #endregion
@@ -429,6 +432,13 @@ namespace Slipstream.Shared
                         OnDefault?.Invoke(this, new EventHandlerArgs<IEvent>(tev));
                     else
                         OnIRacingCommandSendSessionState.Invoke(this, new EventHandlerArgs<Shared.Events.IRacing.IRacingCommandSendSessionState>(tev));
+                    break;
+
+                case Shared.Events.IRacing.IRacingCommandSendRaceFlags tev:
+                    if (OnIRacingCommandSendRaceFlags == null)
+                        OnDefault?.Invoke(this, new EventHandlerArgs<IEvent>(tev));
+                    else
+                        OnIRacingCommandSendRaceFlags.Invoke(this, new EventHandlerArgs<Shared.Events.IRacing.IRacingCommandSendRaceFlags>(tev));
                     break;
 
                 case Shared.Events.IRacing.IRacingDriverIncident tev: